DLL: delete head & tail from list, passes test

+++ doubly_linked_list/doubly_linked_list.py
@@ -70,7 +70,12 @@ class DoublyLinkedList:
     current head's next node the new head of the List.
     Returns the value of the removed Node."""
     def remove_from_head(self):
-        pass
+        # store value of to_be_deleted head
+        value = self.head.value
+        # delete head
+        self.delete(self.head)
+        # return value of deleted node
+        return value
 
     """Wraps the given value in a ListNode and inserts it 
     as the new tail of the list. Don't forget to handle 
@@ -98,7 +103,10 @@ class DoublyLinkedList:
     current tail's previous node the new tail of the List.
     Returns the value of the removed Node."""
     def remove_from_tail(self):
-        pass
+        # opposite of remove from head
+        value = self.tail.value
+        self.delete(self.tail)
+        return value
 
     """Removes the input node from its current spot in the 
     List and inserts it as the new head node of the List."""
@@ -130,12 +138,13 @@ class DoublyLinkedList:
 
         # if tail
         # opposite of head
-        else:
-            # self.tail = self.tail.prev
+        elif node is self.tail:       
+            self.tail = self.tail.prev
             node.delete()
-        
+
         # if regular node, call existing delete function
-        node.delete()
+        else:
+            node.delete()
     """Returns the highest value currently in the list"""
     def get_max(self):
         pass
